Move TaskList into tasklist package
