+++ src/main/java/rakan/Rakan.java
@@ -4,10 +4,12 @@ import rakan.parser.ParsedDeadline;
 import rakan.parser.ParsedEvent;
 import rakan.parser.ParsedMark;
 import rakan.parser.Parser;
+import rakan.storage.Storage;
 import rakan.task.Deadline;
 import rakan.task.Event;
 import rakan.task.Task;
 import rakan.task.ToDo;
+import rakan.tasklist.TaskList;
 import rakan.ui.Ui;
 
 import java.io.IOException;
